move command_strategy and capability_manager in separate classes
