+++ src/hackingBuddyGPT/usecases/__init__.py
@@ -4,5 +4,5 @@ from .web import *
 from .web_api_testing import *
 from .viewer import *
 from .rag import *
-from .command_strategy import *
 from .minimal_linux_privesc import *
+from .linux_privesc import *
