+++ src/hackingBuddyGPT/usecases/minimal_linux_privesc.py
@@ -1,6 +1,6 @@
 from hackingBuddyGPT.capabilities import SSHRunCommand, SSHTestCredential
 from hackingBuddyGPT.usecases.base import use_case
-from hackingBuddyGPT.usecases.command_strategy import CommandStrategy
+from hackingBuddyGPT.strategies import CommandStrategy
 from hackingBuddyGPT.utils import SSHConnection
 
 from mako.template import Template
